+++ metadata.rb
@@ -3,9 +3,8 @@ maintainer 'Corey Hemminger'
 maintainer_email 'hemminger@hotmail.com'
 license 'Apache-2.0'
 description 'Installs/Configures autopatch_ii'
-long_description 'Installs/Configures autopatch_ii'
 version '1.1.0'
-chef_version '>= 12.6' if respond_to?(:chef_version)
+chef_version '>= 12.6'
 
 # The `issues_url` points to the location where issues for this cookbook are
 # tracked.  A `View Issues` link will be displayed on this cookbook's page when
@@ -19,7 +18,7 @@ issues_url 'https://github.com/Stromweld/autopatch_ii/issues'
 #
 source_url 'https://github.com/Stromweld/autopatch_ii'
 
-%w(amazon centos fedora redhat scientific oracle windows).each do |os|
+%w(amazon centos fedora redhat scientific oracle debian ubuntu windows).each do |os|
   supports os
 end